IN-308 refactor: refactor payments table
